Fixes for data loading (#19)
